8/14/15 update
Here's all the sketches from my arduino library.  I think most of them
were already up on git hub.  I also added my libraries folder to the
library folder in the Arduino folder if you ever need to add some of the
libraries in that are required for certain projects.  I also added the
fritzing diagrams with the layouts for the examples we did.  They are in
a folder called Fritz Sketches.